update linux build script, do not try to sign the chages file, ensure that the script is called from its subdirectory

+++ linux/build_lifutils.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
+if [ "`dirname $0`" != "." ] ; then
+   echo "script must be called from its subdirectory"
+   exit 1
+fi
 pushd ..
 rm -f ../lifutils_*.deb
 rm -f ../lifutils_*.changes
-dpkg-buildpackage -b -tc
+dpkg-buildpackage -b -tc -uc
 mv ../lifutils_*.deb linux
 mv ../lifutils_*.changes linux
+rm -f ../lifutils-dbgsym*.deb
+rm -f ../lifutils_*buldinfo
 popd
